Positioned Points number per user on the navbar

+++ BethanysPieShop/Models/ApplicationUser.cs
@@ -8,6 +8,6 @@ namespace QuizIT.Models
 {
     public class ApplicationUser : IdentityUser
     {
-        public long Points { get; set; }
+        public int Points { get; set; }
     }
 }
